+++ CHANGELOG.rst
@@ -19,3 +19,5 @@ Version 0.2
 - Can now pass a list of directories to ensure_dir to be made as part of one path
 - Added num_digits
 - Changed load_json/save_json to allow for splitting large arrays into a .npz file with placeholder keys in the JSON
+- Add gouda.image.polar_to_cartesian
+- Add precision and Matthews correlation coefficient (binary only) to ConfusionMatrix